pyproject.toml
[tool.poetry]
name = "ice"
version = "0.1.0"
description = ""
authors = ["fabio"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
langchain = "^0.2.12"
langchain-openai = "^0.1.20"
langchain-community = "^0.2.11"
langchainhub = "^0.1.20"
langgraph = "^0.1.19"
flask = "^3.0.3"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
